refactor: split enemyHelpers.ts into modular enemy/ directory
Split the 537-line enemyHelpers.ts into focused modules:
- tokenId.ts: Token ID generation and extraction
- piles.ts: Deck/pile initialization (createEnemyTokenPiles)
- drawing.ts: Draw, faction-priority draw, and discard functions
- siteMapping.ts: Site defenders, adventure sites, rampaging colors, visibility
- hexDrawing.ts: Main drawEnemiesForHex orchestration function
- index.ts: Re-exports all public API

Updated all imports across commands and test files to use the new
modular structure at helpers/enemy/index.js.
