+++ packages/core/src/engine/__tests__/pathfindingTerminal.test.ts
@@ -26,7 +26,7 @@ import {
 } from "@mage-knight/shared";
 import { RampagingEnemyType } from "../../types/map.js";
 import type { HexState } from "../../types/map.js";
-import { resetTokenCounter, createEnemyTokenId } from "../helpers/enemyHelpers.js";
+import { resetTokenCounter, createEnemyTokenId } from "../helpers/enemy/index.js";
 import { ENEMY_DIGGERS } from "@mage-knight/shared";
 import { getValidMoveTargets } from "../validActions/movement.js";
