+++ packages/core/src/engine/__tests__/rampagingMovement.test.ts
@@ -22,7 +22,7 @@ import {
 } from "@mage-knight/shared";
 import { RampagingEnemyType } from "../../types/map.js";
 import type { HexState } from "../../types/map.js";
-import { resetTokenCounter, createEnemyTokenId } from "../helpers/enemyHelpers.js";
+import { resetTokenCounter, createEnemyTokenId } from "../helpers/enemy/index.js";
 import { ENEMY_DIGGERS } from "@mage-knight/shared";
 
 describe("Rampaging Enemy Movement Restrictions", () => {
